@@ -45,8 +45,8 @@
     "eslint-plugin-react": "^7.11.0",
     "isomorphic-fetch": "^2.2.1",
     "jest": "^23.6.0",
-    "react": "^16.6.1",
-    "react-dom": "^16.6.1",
+    "react": "^16.7.0-alpha",
+    "react-dom": "^16.7.0-alpha",
     "style-loader": "^0.23.0",
     "webpack": "^4.25.1",
     "webpack-cli": "^3.1.0",
